feat: Support generic input for Parse effect (#1305)
### Problem
Unlike other parsing libraries such as ZIO Parser, Kyo's `Parse` effect
was only made for `Text`/`Char` input and does not support generic
inputs. A common use-case are tokens to separate parsing and lexing.

### Solution
Changed `ParseState` to consume `Chunk[In]` where `In` is the type of
the tokens the parser consumes. The `Parse` is now also parametized with
the input type.

Here is a working example (Scala CLI, put the files in the same
directory) with a JSON parser:
https://gist.github.com/Iltotore/8d4315e1c668645907f6d678c7c202b3

### Notes
- This PR includes breaking changes since `Parse` and some combinators
now require a type parameter.
- Had to do a "dirty trick" with `readAspect` to make it work with
generic input. See its definition and `localReadAspect`'s.
- Despite some methods such as `char` being superseded by their generic
counterpart (e.g `literal[In](expected: In)`), I currently kept them for
backward compatibility and in case we divide `ParseState` into
specialized `TextParseState` and another one for generic chunks.
- Should we make a specialized `ParseState` for `Text` or keep the
current representation of this PR (chunk of `Char`)?
- Should `In` be contravariant in `Parse`?

**DRAFT: Some methods and comments might need to be changed before
marking this PR as ready. I'd still like to receive feedback on its
current state though**

---------

Co-authored-by: Jonathan Winandy <jonathan.winandy@gmail.com>
Co-authored-by: Flavio Brasil <fwbrasil@users.noreply.github.com>