[data] Span improvements (#1401)
I couldn't come up with a new effect this time but this change is in
honor of one of the last start the repo [when it got to
700](https://x.com/fbrasisil/status/1944722949715509346)! Thank you
@evelinemguimaraes 🙏

### Problem

`Span`'s API is still quite limited. I introduced it to avoid boxing in
`Tag` but never got back to expanding the API.

### Solution

Introduce common APIs offered by Scala collections.

### Notes

- I didn't include common collection methods that end up boxing
elements. More details on the scaladoc
- I attempted to make `Span` covariant as suggested by
https://github.com/getkyo/kyo/issues/1185 but it doesn't seem possible
with specialized arrays
- I also expanded most methods to their complete code without calling
other `inline` methods. The goal is to more easily visualize methods
that are too large to inline
- There are methods that are already large but most can be optimized if
we have a specialized `SpanBuilder`. I started working on it but it's
not trivial to ensure it'll specialize. I'll open an issue to follow up.
- I'm also planning to introduce a new `Slice` collection that uses
similar optimization techniques to `Chunk` but is unboxed like `Span`.