Finalizers: introduce add once semantics (#1434)
### Problem
When
[Safepoint](https://github.com/getkyo/kyo/blob/main/kyo-kernel/shared/src/main/scala/kyo/kernel/internal/Safepoint.scala)
`ensure` is called, `ensuring` will re-register the same finalizer on
every loop. Since the finalizer is registered as the `Ensure` class, the
finalizer will only be called atomically once.

### Solution
Instead of removing all instances, we can prevent the addition of
duplicated finalizers. This reduces the size of the ArrayDeque.

Co-authored-by: Jonathan Winandy <jonathan.winandy@gmail.com>